+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] 
+- Added support for reasoning, summarizing errors and import of additional data.
 
 ## [0.1] - 2020-04-14
 - Includes a Validator that can be used to validate RDF data against SHACL files.